Unsubscribe faces from base geometry change
And add that capability to SnapHolder, which now has a new delegate for faces to subscribe to.
SnapHolder gets the base modules present and precalculates a HashSet of their positions, before sending it to the faces. No need to get every module for every face, since it all happens in the same frame!